Move XCTest snapshot helpers to a helper extension separate file
